requirements.txt
altair==4.2.2
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.0
certifi==2020.6.20
cffi @ file:///C:/ci_311/cffi_1676423759166/work
chardet==4.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clearml==1.10.4
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///C:/ci_311/comm_1678376562840/work
contourpy==1.0.7
cryptography @ file:///C:/ci_311/cryptography_1679419210767/work
cycler==0.10.0
debugpy @ file:///C:/ci_311/debugpy_1676426137692/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints==0.4
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
filelock==3.12.0
fonttools==4.39.4
furl==2.1.3
gitdb==4.0.10
GitPython==3.1.31
idna==2.10
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
ipykernel @ file:///C:/ci_311/ipykernel_1678734799670/work
ipython @ file:///D:/bld/ipython_1683226233552/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2 @ file:///C:/ci_311/jinja2_1676424968965/work
jsonschema==4.17.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///D:/bld/jupyter_core_1678994460189/work
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe @ file:///C:/ci_311/markupsafe_1676424152318/work
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.2.1
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
opencv-python==4.7.0.72
orderedmultidict==1.0.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathlib2==2.3.7.post1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1683850015520/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.20.3
psutil @ file:///C:/ci_311_rebuilds/psutil_1679005906571/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==12.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.1b0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
PyJWT==2.4.0
Pympler==1.0.1
pyOpenSSL @ file:///C:/b/abs_de215ipd18/croot/pyopenssl_1678965319166/work
pyparsing==2.4.7
pyrsistent==0.19.3
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
PyYAML==6.0
pyzmq @ file:///C:/b/abs_8b16zbmf46/croot/pyzmq_1682697651374/work
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
requests-toolbelt==1.0.0
rich==13.3.5
roboflow==1.0.8
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.22.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.22.0
sympy @ file:///C:/ci_311_rebuilds/sympy_1679009400182/work
tenacity==8.2.2
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2+cu118
torchvision==0.15.2
tornado @ file:///C:/ci_311/tornado_1676423689414/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.98
urllib3 @ file:///C:/b/abs_3ce53vrdcr/croot/urllib3_1680254693505/work
validators==0.20.0
watchdog==3.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
wget==3.2
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
